+++ torch/csrc/api/include/torch/nn/utils.h
@@ -2,4 +2,4 @@
 
 #include <torch/nn/utils/clip_grad.h>
 #include <torch/nn/utils/convert_parameters.h>
-#include <torch/nn/utils/pad_sequence.h>
+#include <torch/nn/utils/rnn.h>
